+++ corlib/System.IO.Stream.h
@@ -13,8 +13,8 @@ namespace System
         virtual ~Stream();
         virtual int64 Length() = 0;
         virtual int64 Seek(int64, SeekOrigin) = 0;
-        virtual uintptr Position() = 0;
-        virtual void Position(uintptr) = 0;
+        virtual int64 Position() = 0;
+        virtual void Position(int64) = 0;
         virtual bool CanRead() = 0;
         virtual bool CanSeek() = 0;
         virtual bool CanWrite() = 0;
@@ -22,7 +22,7 @@ namespace System
         virtual void Flush() = 0;
         virtual int Read(ByteArray& buffer, int offset, int count) = 0;
         virtual int ReadByte();
-        virtual void SetLength(uintptr) = 0;
+        virtual void SetLength(int64) = 0;
         virtual void Write(ByteArray&, int, int) = 0;
         virtual void WriteByte(byte /*value*/);
       protected:
